+++ djpcms/utils/ajax/__init__.py
@@ -1,2 +1,0 @@
-from jsonbase import *
-from dialog import *