+++ djpcms/views/__init__.py
@@ -1,2 +1,2 @@
-
+from .utils import view_edited_url