+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 Version 0.9.0 - Development
 =======================================
+* Introduced a new Ajax decorator called ``rearrange`` which adds drag & drop functionalities during page editing. 
 * Added abstraction for ``User``, ``Request`` and ``Response`` classes.
 * Added ``save as new`` functionality to :class:`djpcms.views.appview.EditView`.
 * Added `rst` markup handled by sphinx_ in :mod:`djpcms.contrib.flowrepo.markups` module.
@@ -17,7 +18,7 @@ Version 0.9.0 - Development
 * Injecting the instance or ``None`` of current ``url`` into ``request`` object.
 * Added more docs on views and applications.
 * Added :mod:`djpcms.contrib.social` application with ``OAuth``. Still in ``alpha``.
-* **69 unit tests**. Coverage **51%**.
+* **72 unit tests**. Coverage **51%**.
 
 Version 0.8.5 (2010-Nov-16)
 =======================================